Feature/login

+++ src/main/java/com/topicos/backend/controller/AreaController.java
@@ -2,6 +2,8 @@ package com.topicos.backend.controller;
 
 import com.topicos.backend.dto.AreaDTO;
 import com.topicos.backend.dto.request.AreaRequestDTO;
+import com.topicos.backend.exceptions.UnauthorizedException;
+import com.topicos.backend.security.JwtTokenUtil;
 import com.topicos.backend.services.AreaService;
 import java.util.List;
 import lombok.AllArgsConstructor;
@@ -11,6 +13,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
@@ -22,28 +25,39 @@ public class AreaController {
 
   private final AreaService areaService;
 
+  private final JwtTokenUtil jwtTokenUtil;
+
   //CREATE
   @PostMapping("/area/create")
-  public AreaDTO addArea(@RequestBody AreaRequestDTO areaDTO) {
-    return this.areaService.addArea(areaDTO);
+  public AreaDTO addArea(@RequestBody AreaRequestDTO areaDTO, @RequestHeader("Authorization") String token) {
+    if (this.jwtTokenUtil.getAdminFromToken(token)) {
+      return this.areaService.addArea(areaDTO);
+    }
+    throw new UnauthorizedException("The user is not an admin", "The user is not an admin");
   }
 
   //DELETE
   @DeleteMapping("/area/delete")
-  public void deleteArea(@RequestParam Long id) {
-    this.areaService.deleteArea(id);
+  public void deleteArea(@RequestParam Long id, @RequestHeader("Authorization") String token) {
+    if (this.jwtTokenUtil.getAdminFromToken(token)) {
+      this.areaService.deleteArea(id);
+    }
+    throw new UnauthorizedException("The user is not an admin", "The user is not an admin");
   }
 
   //GET
   @GetMapping("/area")
-  public List<AreaDTO> getAllAreas(Long companyId) {
-    return this.areaService.getAllAreas(companyId);
+  public List<AreaDTO> getAllAreas(@RequestHeader("Authorization") String token) {
+    return this.areaService.getAllAreas();
   }
 
-  //MODIFICATE
+  //MODIFICATION
   @PutMapping("/area/modify")
-  public AreaDTO modifyArea(@RequestBody AreaDTO areaDTO) {
-    return this.areaService.modifyArea(areaDTO);
+  public AreaDTO modifyArea(@RequestBody AreaDTO areaDTO, @RequestHeader("Authorization") String token) {
+    if (this.jwtTokenUtil.getAdminFromToken(token)) {
+      return this.areaService.modifyArea(areaDTO);
+    }
+    throw new UnauthorizedException("The user is not an admin", "The user is not an admin");
   }
 
 }
