+++ src/main/java/com/topicos/backend/controller/CompanyController.java
@@ -2,6 +2,8 @@ package com.topicos.backend.controller;
 
 import com.topicos.backend.dto.CompanyDTO;
 import com.topicos.backend.dto.request.CompanyRequestDTO;
+import com.topicos.backend.exceptions.UnauthorizedException;
+import com.topicos.backend.security.JwtTokenUtil;
 import com.topicos.backend.services.CompanyService;
 import java.util.List;
 import lombok.AllArgsConstructor;
@@ -11,6 +13,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
@@ -22,27 +25,35 @@ public class CompanyController {
 
   private final CompanyService companyService;
 
+  private final JwtTokenUtil jwtTokenUtil;
+
   //CREATE
   @PostMapping("/company/create")
-  public CompanyDTO addCompany(@RequestBody CompanyRequestDTO company) {
-    return this.companyService.addCompany(company);
+  public CompanyDTO addCompany(@RequestBody CompanyRequestDTO company, @RequestHeader("Authorization") String token) {
+    if (this.jwtTokenUtil.getAdminFromToken(token)) {
+      return this.companyService.addCompany(company);
+    }
+    throw new UnauthorizedException("The user is not an admin", "The user is not an admin");
   }
 
   //DELETE
   @DeleteMapping("/company/delete")
-  public void deleteCompany(@RequestParam Long id) {
-    this.companyService.deleteCompany(id);
+  public void deleteCompany(@RequestParam Long id, @RequestHeader("Authorization") String token) {
+    if (this.jwtTokenUtil.getAdminFromToken(token)) {
+      this.companyService.deleteCompany(id);
+    }
+    throw new UnauthorizedException("The user is not an admin", "The user is not an admin");
   }
 
   //GET
   @GetMapping("/company")
-  public List<CompanyDTO> getAllCompanys() {
+  public List<CompanyDTO> getAllCompanies(@RequestHeader("Authorization") String token) {
     return this.companyService.getAllCompanies();
   }
 
-  //MODIFICATE
+  //MODIFICATION
   @PutMapping("/company/modify")
-  public CompanyDTO modifyCompany(@RequestBody CompanyDTO company) {
+  public CompanyDTO modifyCompany(@RequestBody CompanyDTO company, @RequestHeader("Authorization") String token) {
     return this.companyService.modifyCompany(company);
   }
